requirements.txt
boto==2.49.0
click==8.1.3
Flask==2.1.2
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
SQLAlchemy==1.4.37
Werkzeug==2.1.2